Fix issues with 24b slicing (#2536)
* Handle 24b slice lists

Ensure that if possible, 24b slice lists (those by themselves) are split at slice boundaries. Need
to add code to split 24b slice lists created out of earlier slicing pass.

Resolve slicing issues with comp_868.p4

P4C-1244 #assign #start-work #fixed Handle 24b slice lists #fixed

* Clean up code and add comments

P4C-1244 #assign #start-work #fixed Clean up code and add comments #fixed

+++ bf-p4c/phv/action_phv_constraints.cpp
@@ -769,7 +769,7 @@ bool ActionPhvConstraints::parser_constant_extract_satisfied(
             }
             // Need to pack within 4 consecutive bits for 16b fields and within 3 consecutive bits
             // for 32b fields.
-            int maxAllowed = c.is(PHV::Size::b16) ? 4 : 3;
+            size_t maxAllowed = c.is(PHV::Size::b16) ? 4 : 3;
             if (count <= maxAllowed) continue;
             LOG5("\t\t\tIn container " << c << ", POV slices " << slice << " and " << slice1 <<
                  " are packed " << count << " (maxAllowed: " << maxAllowed << "b) bits apart." <<
